// src/app/shared/models/Users.ts
export class Users {
  firstName: string | undefined;
  lastName: string | undefined;
  dateOfBirth: Date | undefined;
  email: string | undefined;
  password: string | undefined;
  phone: number | undefined;
  address: string | undefined;
  city: string | undefined;
  pincode: string | undefined;
  userRole: string | undefined;
  user_Id: string | undefined;
  RowStatus:string = "UPDATED";
}